Add support for ZIP standard cp437 encoding
This encoding is used by ZIP standard. We fallback to UTF-8 encoding if CP437 encoding is unavailable on the platform or it is detected that the text is actually in UTF-8 encoding. The latter is needed because files created by Info-ZIP with UTF-8 names doesn't actually have bit flag which is required (again, according to standard) for UTF-8.